requirements.txt
torch
transformers
datasets
peft
accelerate
scikit-learn
pyyaml
tqdm
pandas
matplotlib
